Refactored get-token, token
get-token now takes in a c-style string to parse instead of assuming
std::input. I was originally going to make it a std::istream, but then I
remembered that the llvm coding standard doesn't allow the use of
istream. I looked it up, and they said to use llvm::MemoryBuffer.
Looking at the interface, requiring a c-style string instead services
the people using a MemoryBuffer (as MemoryBuffer guarnatees a trailing
NULL in all situations) while adding flexibility (the test suite can pass
in literal strings, for example).

token had two changes made to it. The first is that all of the Tags lost
their tok_ prefix, and exter gained a k_ prefix because extern is a
keyword in c++ (k for kaleidoscope extern, as oppossed to c++ extern).
Additionally, Comment struct was added as I believe it is better to
tokenize comments like everything else for the purpose of, for example,
debugging.